indian-trading-bot/requirements.txt
vectorbt>=0.25.0
psycopg2-binary>=2.9.0
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0
python-dotenv>=0.19.0
sqlalchemy>=1.4.0
dhanhq>=1.3.0
yfinance>=0.2.0
click>=8.0.0